tests: refactor tests related to functions removed from ProviderThread class
