+++ tests/fed_reg/test_merge_components.py
@@ -8,59 +8,18 @@ from fed_reg.provider.schemas_extended import (
     ProjectCreate,
     RegionCreateExtended,
 )
-from pytest_cases import parametrize_with_cases
 
-from src.models.identity_provider import Issuer
-from src.models.provider import Kubernetes, Openstack
-from src.providers.core import ProviderThread
-from tests.schemas.utils import (
-    auth_method_dict,
-    issuer_dict,
-    openstack_dict,
-    project_dict,
-    sla_dict,
-    user_group_dict,
-)
+from src.utils import merge_components
 from tests.utils import random_lower_string, random_url
 
 
-class CaseProviderThread:
-    def case_openstack(self) -> ProviderThread:
-        provider = Openstack(
-            **openstack_dict(),
-            identity_providers=[auth_method_dict()],
-            projects=[project_dict()],
-        )
-        issuer = Issuer(
-            **issuer_dict(),
-            token=random_lower_string(),
-            user_groups=[{**user_group_dict(), "slas": [sla_dict()]}],
-        )
-        return ProviderThread(provider_conf=provider, issuers=[issuer])
-
-    def case_k8s(self) -> ProviderThread:
-        provider = Kubernetes(
-            **openstack_dict(),
-            identity_providers=[auth_method_dict()],
-            projects=[project_dict()],
-        )
-        issuer = Issuer(
-            **issuer_dict(),
-            token=random_lower_string(),
-            user_groups=[{**user_group_dict(), "slas": [sla_dict()]}],
-        )
-        return ProviderThread(provider_conf=provider, issuers=[issuer])
-
-
-@parametrize_with_cases("provider_thread_item", cases=CaseProviderThread)
-def test_merge_data_single_entry(
-    provider_thread_item: ProviderThread,
+def test_merge_components_single_entry(
     identity_provider_create: IdentityProviderCreateExtended,
     project_create: ProjectCreate,
     region_create: RegionCreateExtended,
 ):
     siblings = [(identity_provider_create, project_create, region_create)]
-    identity_providers, projects, regions = provider_thread_item.merge_data(siblings)
+    identity_providers, projects, regions = merge_components(siblings)
     assert len(identity_providers) == 1
     assert isinstance(identity_providers[0], IdentityProviderCreateExtended)
     assert len(projects) == 1
@@ -69,9 +28,7 @@ def test_merge_data_single_entry(
     assert isinstance(regions[0], RegionCreateExtended)
 
 
-@parametrize_with_cases("provider_thread_item", cases=CaseProviderThread)
-def test_merge_data_multi_entries(
-    provider_thread_item: ProviderThread,
+def test_merge_components_multi_entries(
     identity_provider_create: IdentityProviderCreateExtended,
     project_create: ProjectCreate,
     region_create: RegionCreateExtended,
@@ -86,15 +43,13 @@ def test_merge_data_multi_entries(
         (identity_provider_create, project_create, region_create),
         (idp2, proj2, reg2),
     ]
-    identity_providers, projects, regions = provider_thread_item.merge_data(siblings)
+    identity_providers, projects, regions = merge_components(siblings)
     assert len(identity_providers) == 2
     assert len(projects) == 2
     assert len(regions) == 2
 
 
-@parametrize_with_cases("provider_thread_item", cases=CaseProviderThread)
-def test_merge_data_multi_entries_to_merge(
-    provider_thread_item: ProviderThread,
+def test_merge_components_multi_entries_to_merge(
     identity_provider_create: IdentityProviderCreateExtended,
     project_create: ProjectCreate,
     region_create: RegionCreateExtended,
@@ -106,15 +61,13 @@ def test_merge_data_multi_entries_to_merge(
         (identity_provider_create, project_create, region_create),
         (idp2, proj2, reg2),
     ]
-    identity_providers, projects, regions = provider_thread_item.merge_data(siblings)
+    identity_providers, projects, regions = merge_components(siblings)
     assert len(identity_providers) == 1
     assert len(projects) == 1
     assert len(regions) == 1
 
 
-@parametrize_with_cases("provider_thread_item", cases=CaseProviderThread)
-def test_merge_data_multi_entries_filter_projects(
-    provider_thread_item: ProviderThread,
+def test_merge_components_multi_entries_filter_projects(
     identity_provider_create: IdentityProviderCreateExtended,
     project_create: ProjectCreate,
     region_create: RegionCreateExtended,
@@ -137,7 +90,7 @@ def test_merge_data_multi_entries_filter_projects(
     )
 
     siblings = [(identity_provider_create, project_create, region_create)]
-    items = provider_thread_item.merge_data(siblings)
+    items = merge_components(siblings)
     regions = items[2]
     assert len(regions) == 1
     assert len(regions[0].compute_services[0].flavors[0].projects) == 1
